database: Refactor Loan Dialog
